Move unix specific part in its own implementation
